server/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
numpy>=1.26.4
pydantic>=2.7.0
starlette>=0.37.2
requests>=2.32.0